Separated FAQ - Child files & Child edit clas to function component
